Rename "eigency_typedefs.h" to just "eigency.h". Update README.
